Add option to include details in Memcache constructor
This commit adds a new parameter to the `Memcache` class constructor, allowing the inclusion of details. The new parameter, `$details`, is set to `false` by default. When set to `true`, it enables the generation of detailed information for groups.

The change introduces a static property, `details`, to the `Memcache` class, and assigns the value of `$details` to it within the constructor. This property will be used to control whether group details are processed or not.

The code modification includes the following:
- Added a new static boolean property, `details`, to the `Memcache` class.
- Modified the constructor to accept an optional `$details` parameter.
- Set the value of `self::$details` to `$details` within the constructor.
- Modified the code block that generates group details to conditionally execute based on the value of `self::$details`.
- Updated the logging output to include group details when `self::$details` is `true`.

These changes provide more flexibility in the `Memcache` class, allowing developers to choose whether they want to include detailed information about groups.